# *: optimize external iters for single-file, no-rangekey cases