This change updates NewExternalIter to skip creation of a point-key
mergingIter if there is only one merging iter level and there are
no range dels. It also skips creation of a simpleLevelIter if there
is only one file that would be contained in it. And it skips
creation of a rangeKey interleaving iterator if there are no range
keys. These optimizations avoid overhead that's only necessary for
multi-file, range-key cases.

Together, these optimizations produce a ~45% performance gain
for forward iteration:

```
name                                                   old time/op  new time/op  delta
SSTIterator/keys=10000/variant=pebble/verify=false-10  1.22ms ± 0%  0.67ms ± 0%  -45.27%  (p=0.001 n=5+10)
```