@itsbilal 2022-08-22
The same benchmark before this change (old = legacy iter, new = new iter before this change):

```
name                                                   old time/op  new time/op  delta
SSTIterator/keys=10000/variant=pebble/verify=false-10   660µs ± 1%  1224µs ± 0%  +85.54%  (p=0.008 n=5+5)
```

Vs after this change (old is still the legacy iterator)
```
name                                                   old time/op  new time/op  delta
SSTIterator/keys=10000/variant=pebble/verify=false-10   660µs ± 1%   990µs ± 0%  +50.15%  (p=0.008 n=5+5)
```

There's still a noticeable performance overhead and while we'll not be able to get to 0 because we need the toplevel `Iterator` for figuring out range key presence/overlaps etc, there are possibly other things I can keep digging into after this change.